Разбить файл тестов на файлы и папки, сделать Node класс внутренним и приватным, удалить clear из интерфейса HistoryManager
