+++ src/ru/alexgur/kanban/service/HistoryManager.java
@@ -8,8 +8,6 @@ public interface HistoryManager {
 
     void add(Task task);
 
-    void clear();
-
     void remove(int id);
 
     List<Task> getHistory();
